Move all errors to one common file
