+++ src/lib.rs
@@ -1,11 +1,12 @@
 pub mod connection;
 mod entity;
-pub use crate::entity::{DatastoreEntity, DatastoreProperties, DatastoreValue, DatastoreParseError, DatastoreEntityCollection, ResultCollection};
+pub mod error;
+pub use crate::entity::{DatastoreEntity, DatastoreProperties, DatastoreValue, DatastoreEntityCollection, ResultCollection};
 pub use crate::connection::{DatastoreConnection};
+pub use crate::error::{DatastoreParseError, DatastoreClientError, DatastorersError};
 
 pub use datastore_entity_derives::DatastoreManaged;
 
-use thiserror::Error;
 use google_datastore1::schemas::{
     BeginTransactionRequest, BeginTransactionResponse, CommitRequest, CommitResponse, Entity,
     Filter, Key, KindExpression, LookupRequest, LookupResponse, Mutation, PathElement,
@@ -17,34 +18,6 @@ use std::convert::TryInto;
 use std::convert::TryFrom;
 
 const DEFAULT_PAGE_SIZE: i32 = 50;
-
-#[derive(Error, Debug, PartialEq)]
-pub enum DatastoreClientError {
-    #[error("entity not found")]
-    NotFound,
-    #[error("multiple entities found, single result expected")]
-    AmbigiousResult,
-    #[error("missing key, entity cannot be commited")]
-    KeyMissing,
-    #[error("failed to assign key to inserted entity")]
-    KeyAssignmentFailed,
-    #[error("delete operation failed")]
-    DeleteFailed,
-    #[error("unexpected response data")]
-    ApiDataError,
-    #[error("no more pages to fetch")]
-    NoMorePages,
-}
-
-#[derive(Error, Debug)]
-pub enum DatastorersError {
-    #[error(transparent)]
-    ParseError(#[from] DatastoreParseError),
-    #[error(transparent)]
-    DatastoreError(#[from] google_datastore1::Error),
-    #[error(transparent)]
-    DatastoreClientError(#[from] DatastoreClientError)
-}
 
 
 pub fn get_one_by_id(
